refactor(memory): share provenance mutation wrapper
